modified:   gfuncs.c 	modified:   gfuncs.h 	modified:   pfuncs.c 	modified:   pfuncs.h 	modified:   ui 	modified:   ui.c Program now displays text without breaking up words over lines. Also it seems as though the memory leaks are fixed, barring a pesky strange one with the subject array. Perhaps a pointery adjustment would fix this.
Overall though, the project is a bit of a mess. As I am also getting to
the point where I will make the code public, a tidying up/reorganisation
might improve the presentation/readability of my code, and help me to
think about how to lay things out for future projects.